porting fs_cache tests v1
