+++ cache/fs_cache_test.go
@@ -39,75 +39,6 @@ func TestWriteReadHeader(t *testing.T) {
 	}
 }
 
-func TestKeyString(t *testing.T) {
-	testCases := []struct {
-		key      *Key
-		expected string
-	}{
-		{
-			key: &Key{
-				Query: []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-			},
-			expected: "bebe3382e36ffdeea479b45d827b208a",
-		},
-		{
-			key: &Key{
-				Query:          []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-				AcceptEncoding: "gzip",
-			},
-			expected: "498c1af30fb94280fd7c7225c0c8fb39",
-		},
-		{
-			key: &Key{
-				Query:          []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-				AcceptEncoding: "gzip",
-				DefaultFormat:  "JSON",
-			},
-			expected: "720292aa0647cc5e53e0b6e6033eef34",
-		},
-		{
-			key: &Key{
-				Query:          []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-				AcceptEncoding: "gzip",
-				DefaultFormat:  "JSON",
-				Database:       "foobar",
-			},
-			expected: "5c6a70736d71e570faca739c4557780c",
-		},
-		{
-			key: &Key{
-				Query:          []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-				AcceptEncoding: "gzip",
-				DefaultFormat:  "JSON",
-				Database:       "foobar",
-				Namespace:      "ns123",
-			},
-			expected: "08b4baf6825e53bbd18136a88abda4f8",
-		},
-		{
-			key: &Key{
-				Query:          []byte("SELECT 1 FROM system.numbers LIMIT 10"),
-				AcceptEncoding: "gzip",
-				DefaultFormat:  "JSON",
-				Database:       "foobar",
-				Compress:       "1",
-				Namespace:      "ns123",
-			},
-			expected: "0e043f23ccd1b9039b33623b3b7c114a",
-		},
-	}
-
-	for _, tc := range testCases {
-		s := tc.key.String()
-		if !cachefileRegexp.MatchString(s) {
-			t.Fatalf("invalid key string format: %q", s)
-		}
-		if s != tc.expected {
-			t.Fatalf("unexpected key string: %q; expecting: %q", s, tc.expected)
-		}
-	}
-}
-
 func TestCacheClose(t *testing.T) {
 	for i := 0; i < 10; i++ {
 		c := newTestCache(t)
@@ -124,7 +55,7 @@ func TestCacheAddGet(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d", i)),
 		}
 		trw := &testResponseWriter{}
-		crw, err := NewTmpFileResponseWriter(trw, "/dir")
+		crw, err := NewTmpFileResponseWriter(trw, "/tmp")
 		if err != nil {
 			t.Fatalf("cannot create response writer: %s", err)
 		}
@@ -139,7 +70,17 @@ func TestCacheAddGet(t *testing.T) {
 		if _, err := io.Copy(crw, bs); err != nil {
 			t.Fatalf("cannot send response to cache: %s", err)
 		}
-		if err := crw.Commit(); err != nil {
+
+		f, err := crw.GetFile()
+		if err != nil {
+			t.Fatalf("cannot get file: %s", err)
+		}
+
+		if _, err := c.Put(f, key); err != nil {
+			t.Fatalf("failed to put it to cache: %s", err)
+		}
+
+		if err := SendResponseFromFile(trw, f, 0*time.Second, http.StatusOK); err != nil {
 			t.Fatalf("cannot commit response to cache: %s", err)
 		}
 
@@ -170,7 +111,7 @@ func TestCacheAddGet(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d", i)),
 		}
 		trw := &testResponseWriter{}
-		if err := c.WriteTo(trw, key); err != nil {
+		if err := c.Get(trw, key); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
 		value := fmt.Sprintf("value %d", i)
@@ -205,7 +146,7 @@ func TestCacheAddGet(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d", i)),
 		}
 		trw := &testResponseWriter{}
-		if err := c1.WriteTo(trw, key); err != nil {
+		if err := c1.Get(trw, key); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
 		value := fmt.Sprintf("value %d", i)
@@ -241,7 +182,7 @@ func TestCacheMiss(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d cache miss", i)),
 		}
 		trw := &testResponseWriter{}
-		err := c.WriteTo(trw, key)
+		err := c.Get(trw, key)
 		if err == nil {
 			t.Fatalf("expecting error")
 		}
@@ -251,92 +192,89 @@ func TestCacheMiss(t *testing.T) {
 	}
 }
 
-func TestPendingEntries(t *testing.T) {
-	cfg := config.Cache{
-		Name:      "foobar",
-		Dir:       testDir,
-		MaxSize:   1e6,
-		Expire:    config.Duration(time.Minute),
-		GraceTime: config.Duration(30 * time.Second),
-	}
-	c, err := New(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer c.Close()
-
-	key := &Key{
-		Query: []byte("SELECT pending entries"),
-	}
-	value := "value for pending entries"
-
-	trw := &testResponseWriter{}
-	err = c.WriteTo(trw, key)
-	if err == nil {
-		t.Fatalf("expecting error")
-	}
-	if err != ErrMissing {
-		t.Fatalf("unexpected error: %s; expecting %s", err, ErrMissing)
-	}
-
-	ch := make(chan error)
-	go func() {
-		trw := &testResponseWriter{}
-		// This should be delayed until the main goroutine writes
-		// the value for the given key.
-		if err := c.WriteTo(trw, key); err != nil {
-			ch <- fmt.Errorf("error in WriteTo: %s", err)
-			return
-		}
-		if string(trw.b) != value {
-			ch <- fmt.Errorf("unexpected response sent to client: %q; expecting %q", trw.b, value)
-			return
-		}
-		ch <- nil
-	}()
-
-	// Verify that the started goroutine is blocked.
-	select {
-	case err := <-ch:
-		t.Fatalf("the goroutine prematurely finished and returned err: %v", err)
-	case <-time.After(200 * time.Millisecond):
-	}
-
-	// Write the value to the cache.
-	trw = &testResponseWriter{}
-	crw, err := c.NewResponseWriter(trw, key)
-	if err != nil {
-		t.Fatalf("cannot create response writer: %s", err)
-	}
-
-	bs := bytes.NewBufferString(value)
-	if _, err := io.Copy(crw, bs); err != nil {
-		t.Fatalf("cannot send response to cache: %s", err)
-	}
-	if err := crw.Commit(); err != nil {
-		t.Fatalf("cannot commit response to cache: %s", err)
-	}
-
-	// Verify that the started goroutine exits.
-	select {
-	case <-time.After(3 * time.Second):
-	case err := <-ch:
-		if err != nil {
-			t.Fatalf("unexpected error returned from the goroutine: %s", err)
-		}
-	}
-}
+//func TestPendingEntries(t *testing.T) {
+//	cfg := config.Cache{
+//		Name:      "foobar",
+//		Dir:       testDir,
+//		MaxSize:   1e6,
+//		Expire:    config.Duration(time.Minute),
+//		GraceTime: config.Duration(30 * time.Second),
+//	}
+//	c, err := New(cfg)
+//	if err != nil {
+//		t.Fatal(err)
+//	}
+//	defer c.Close()
+//
+//	key := &Key{
+//		Query: []byte("SELECT pending entries"),
+//	}
+//	value := "value for pending entries"
+//
+//	trw := &testResponseWriter{}
+//	err = c.WriteTo(trw, key)
+//	if err == nil {
+//		t.Fatalf("expecting error")
+//	}
+//	if err != ErrMissing {
+//		t.Fatalf("unexpected error: %s; expecting %s", err, ErrMissing)
+//	}
+//
+//	ch := make(chan error)
+//	go func() {
+//		trw := &testResponseWriter{}
+//		// This should be delayed until the main goroutine writes
+//		// the value for the given key.
+//		if err := c.WriteTo(trw, key); err != nil {
+//			ch <- fmt.Errorf("error in WriteTo: %s", err)
+//			return
+//		}
+//		if string(trw.b) != value {
+//			ch <- fmt.Errorf("unexpected response sent to client: %q; expecting %q", trw.b, value)
+//			return
+//		}
+//		ch <- nil
+//	}()
+//
+//	// Verify that the started goroutine is blocked.
+//	select {
+//	case err := <-ch:
+//		t.Fatalf("the goroutine prematurely finished and returned err: %v", err)
+//	case <-time.After(200 * time.Millisecond):
+//	}
+//
+//	// Write the value to the cache.
+//	trw = &testResponseWriter{}
+//	crw, err := c.NewResponseWriter(trw, key)
+//	if err != nil {
+//		t.Fatalf("cannot create response writer: %s", err)
+//	}
+//
+//	bs := bytes.NewBufferString(value)
+//	if _, err := io.Copy(crw, bs); err != nil {
+//		t.Fatalf("cannot send response to cache: %s", err)
+//	}
+//	if err := crw.Commit(); err != nil {
+//		t.Fatalf("cannot commit response to cache: %s", err)
+//	}
+//
+//	// Verify that the started goroutine exits.
+//	select {
+//	case <-time.After(3 * time.Second):
+//	case err := <-ch:
+//		if err != nil {
+//			t.Fatalf("unexpected error returned from the goroutine: %s", err)
+//		}
+//	}
+//}
 
 func TestCacheRollback(t *testing.T) {
 	c := newTestCache(t)
 	defer c.Close()
 
 	for i := 0; i < 10; i++ {
-		key := &Key{
-			Query: []byte(fmt.Sprintf("SELECT %d cache rollback", i)),
-		}
 		trw := &testResponseWriter{}
-		crw, err := c.NewResponseWriter(trw, key)
+		crw, err := NewTmpFileResponseWriter(trw, "/tmp")
 		if err != nil {
 			t.Fatalf("cannot create response writer: %s", err)
 		}
@@ -346,7 +284,11 @@ func TestCacheRollback(t *testing.T) {
 		if _, err := io.Copy(crw, bs); err != nil {
 			t.Fatalf("cannot send response to cache: %s", err)
 		}
-		if err := crw.Rollback(); err != nil {
+		if f, err := crw.GetFile(); err == nil {
+			if err := SendResponseFromFile(trw, f, 0*time.Second, http.StatusOK); err != nil {
+				t.Fatalf("cannot commit response to cache: %s", err)
+			}
+		} else {
 			t.Fatalf("cannot rollback response: %s", err)
 		}
 
@@ -362,7 +304,7 @@ func TestCacheRollback(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d cache rollback", i)),
 		}
 		trw := &testResponseWriter{}
-		err := c.WriteTo(trw, key)
+		err := c.Get(trw, key)
 		if err == nil {
 			t.Fatalf("expecting non-nil error")
 		}
@@ -379,7 +321,7 @@ func TestCacheClean(t *testing.T) {
 		MaxSize: 8192,
 		Expire:  config.Duration(time.Minute),
 	}
-	c, err := New(cfg)
+	c, err := newFSCache(cfg, 0 * time.Second)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -391,7 +333,7 @@ func TestCacheClean(t *testing.T) {
 			Query: []byte(fmt.Sprintf("SELECT %d cache clean", i)),
 		}
 		trw := &testResponseWriter{}
-		crw, err := c.NewResponseWriter(trw, key)
+		crw, err := NewTmpFileResponseWriter(trw, "/tmp")
 		if err != nil {
 			t.Fatalf("cannot create response writer: %s", err)
 		}
@@ -401,9 +343,16 @@ func TestCacheClean(t *testing.T) {
 		if _, err := io.Copy(crw, bs); err != nil {
 			t.Fatalf("cannot send response to cache: %s", err)
 		}
-		if err := crw.Commit(); err != nil {
-			t.Fatalf("cannot commit response to cache: %s", err)
+
+		f, err := crw.GetFile()
+		if err != nil {
+			t.Fatalf("cannot get file: %s", err)
 		}
+
+		if _, err := c.Put(f, key); err != nil {
+			t.Fatalf("failed to put it to cache: %s", err)
+		}
+		crw.Close()
 	}
 
 	// Forcibly clean the cache
@@ -445,7 +394,7 @@ func (trw *testResponseWriter) Header() http.Header {
 
 func (trw *testResponseWriter) WriteHeader(statusCode int) {}
 
-func newTestCache(t *testing.T) *FSCache {
+func newTestCache(t *testing.T) *FileSystemCache {
 	t.Helper()
 
 	cfg := config.Cache{
@@ -454,7 +403,7 @@ func newTestCache(t *testing.T) *FSCache {
 		MaxSize: 1e6,
 		Expire:  config.Duration(time.Minute),
 	}
-	c, err := New(cfg)
+	c, err := newFSCache(cfg, time.Second)
 	if err != nil {
 		t.Fatal(err)
 	}
